relocate console helpers
