+++ cstar/cli/workplan/run.py
@@ -18,6 +18,7 @@ from cstar.base.utils import slugify
 from cstar.cli.common import (
     cb_pipeline,
     normalize_runid,
+    present,
     set_env,
     set_flag,
     update_loggers,
@@ -27,7 +28,6 @@ from cstar.cli.workplan.shared import (
     colored,
     console,
     list_runs,
-    present,
 )
 from cstar.entrypoint.utils import (
     ARG_CLOBBER,
